fix: cleanup slop from local-mode onboarding implementation
